requirements.txt
python==3.9.16
pip==23.0.1
django==4.1
djangorestframework==3.14.0
psycopg2==2.9.6
django-cors-headers==4.0.0
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2